# docs/development/api/openapi.json
{
  "components": {
    "schemas": {
      "ErrorBody": {
        "description": "Standard error payload returned by every failed request.\n\nThe envelope is uniform across every error the server emits, including\nrouting fallbacks and extractor rejections (`API-039`).",
        "properties": {
          "_links": {
            "$ref": "#/components/schemas/SelfLinks",
            "description": "Link to the request that produced the error."
          },
          "error": {
            "description": "Human-readable description of the error.",
            "example": "An error message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "_links"
        ],
        "type": "object"
      },
...
      "Link": {
        "description": "A single HAL link.",
        "properties": {
          "href": {
            "description": "Target URI reference of the link.",
            "type": "string"
          },
          "templated": {
            "description": "Whether `href` is a URI template, per RFC 6570.",
            "example": true,
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "required": [
          "href"
        ],
        "type": "object"
      },
...
      "PatchCredentialRequest": {
        "description": "Payload to change the password behind a credential.",
        "properties": {
          "password": {
            "description": "New password; never returned by the API.\n\nMust be at least 8 characters long and contain at least one symbol\n(a non-alphanumeric character).",
            "minLength": 8,
            "pattern": "[^A-Za-z0-9]",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "password"
        ],
        "type": "object"
      },
...
      "SelfLinks": {
        "description": "The `self` link of a HAL representation.\n\nEvery resource representation carries the URI it was served from, so clients\nnever construct or hardcode it themselves (`API-032`, `API-037`).",
        "properties": {
          "self": {
            "$ref": "#/components/schemas/Link",
            "description": "Link to the resource itself."
          }
        },
        "required": [
          "self"
        ],
        "type": "object"
      },
...
      "i64": {
        "format": "int64",
        "type": "integer"
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API of the Mnemorium service",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Mnemorium API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/identity/credential/{id}": {
      "patch": {
        "description": "Only the Root Admin (user identifier `0`, role `Admin`) may invoke this\nendpoint; any other authenticated caller is rejected with `403`. The Root\nAdmin may change its own credential as well as the credential of any other\nuser. The new password must satisfy the password policy: at least 8\ncharacters long and containing at least one symbol.",
        "operationId": "patch_credential",
        "parameters": [
          {
            "description": "Identifier of the credential whose password is changed",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/i64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid payload or password policy violation"
          },
          "401": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Caller is not the Root Admin"
          },
          "404": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unknown credential"
          },
          "500": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Change the password behind a credential",
        "tags": [
          "identity"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://0.0.0.0:4080/api/v1"
    }
  ],
...
}